fix: 调整项目的收费项计算方法

+++ yudao-module-jl/yudao-module-jl-biz/src/main/java/cn/iocoder/yudao/module/jl/controller/admin/project/vo/ProjectCostStatsRespVO.java
@@ -25,7 +25,7 @@ public class ProjectCostStatsRespVO {
     private Long procurementCost;
 
     @Schema(description = "收费项的成本")
-    private Long chargeItemCost;
+    private BigDecimal chargeItemCost;
 
     @Schema(description = "委外的成本")
     private Long outsourceCost;
